Fix for subtraction bug
VF got set incorrectly when subtracting (sub and subn).
This should fix a lot of issues you have with games.

Found this when porting this to the TI-Nspire. Games such as Brix didn't work properly, and after a lot of searching and decompiling Brix I figured out that this was the problem. 
I made multiple other changes, but those are rather optimizations, so I'm not going to push that. This is needed for good emulator functioning.